backend/requirements.txt
flask
flask-cors
load_dotenv
